refactor: split deployments.ts into command handler and shared helpers
