+++ src/command-dispatch.ts
@@ -10,6 +10,7 @@
 
 import type { AnyCommandHandler } from "./command-framework.ts";
 import { completionCommand } from "./commands/completion.ts";
+import { deployCommand } from "./commands/deploy.ts";
 import { getFormCommand } from "./commands/forms.ts";
 import {
 	assignFallbackCommand,
@@ -119,7 +120,6 @@ import {
 } from "./commands/user-tasks.ts";
 import { setVariableCommand } from "./commands/variables.ts";
 import { watchCommand } from "./commands/watch.ts";
-import { deployCommand } from "./deployments.ts";
 
 /**
  * Dispatch map keyed by "verb:resource".
