+++ src/commands/watch.ts
@@ -5,7 +5,10 @@
 import { existsSync, realpathSync, statSync, watch } from "node:fs";
 import { basename, extname, resolve } from "node:path";
 import { defineCommand } from "../command-framework.ts";
-import { deployResources, findProcessApplicationRoot } from "../deployments.ts";
+import {
+	deployResources,
+	findProcessApplicationRoot,
+} from "../deploy-helpers.ts";
 import { normalizeToError } from "../errors.ts";
 import { isIgnored, loadIgnoreRules, resolveIgnoreBaseDir } from "../ignore.ts";
 import { DEPLOY_COOLDOWN } from "../watch-constants.ts";
